requirements.txt
beautifulsoup4==4.12.3
lxml==5.1.0
requests==2.31.0
python-dateutil==2.8.2
langdetect==1.0.9
Flask==3.0.2
flask-cors==4.0.0
SQLAlchemy==2.0.27
schedule==1.2.1
tqdm==4.66.ls
flask_cors==4.0.0
